refactor(live-common): relocate FF domain helpers out of featureFlags module (LIVE-31640)
Move the flag-reading domain helpers into their domain modules and off the
types-live feature types, repointing both apps' importers:
- featureFlags/stakePrograms -> earn/stakePrograms
- featureFlags/helper.isRecoverDisplayed -> recover/isRecoverDisplayed
- featureFlags/remoteABTesting -> analytics/remoteABTesting

Align the app analytics boundary to @shared: type analyticsFeatureFlagMethod
with @shared FeatureId/Features in both apps, and migrate the desktop/mobile
useVersionedStakePrograms hooks to Features["stakePrograms"], dropping the
types-live brand casts (the relocated stakeProgram helpers now consume @shared).
