+++ apps/ledger-live-desktop/src/renderer/analytics/segment.ts
@@ -4,15 +4,14 @@ import {
   getStablecoinYieldSetting,
   getBitcoinYieldSetting,
   getEthDepositScreenSetting,
-} from "@ledgerhq/live-common/featureFlags/stakePrograms/index";
+} from "@ledgerhq/live-common/earn/stakePrograms/index";
 import { runOnceWhen } from "@ledgerhq/live-common/utils/runOnceWhen";
 import { LiveConfig } from "@ledgerhq/live-config/LiveConfig";
 import { getEnv } from "@ledgerhq/live-env";
 import { getDefaultAccountName } from "@ledgerhq/live-wallet/accountName";
-// `Feature, FeatureId, Features` stay on types-live: `analyticsFeatureFlagMethod` is set by
-// FirebaseFeatureFlagsProvider whose prop type uses live-common's `GetFeature` shape (LIVE-31228).
-import type { AccountLike, Feature, FeatureId, Features } from "@ledgerhq/types-live";
+import type { AccountLike } from "@ledgerhq/types-live";
 import { idsToLanguage } from "@ledgerhq/types-live";
+import type { Feature, FeatureId, Features } from "@shared/feature-flags";
 import invariant from "invariant";
 import { useCallback, useContext } from "react";
 import type * as Redux from "redux";
