+++ apps/ledger-live-desktop/src/renderer/components/Onboarding/Screens/SelectUseCase/index.tsx
@@ -4,7 +4,7 @@ import { useTranslation, Trans } from "react-i18next";
 import { Flex, Text } from "@ledgerhq/react-ui";
 import { DeviceModelId } from "@ledgerhq/devices";
 import { useFeature } from "@features/platform-feature-flags";
-import { isRecoverDisplayed } from "@ledgerhq/live-common/featureFlags/index";
+import { isRecoverDisplayed } from "@ledgerhq/live-common/recover/isRecoverDisplayed";
 import { useDispatch } from "LLD/hooks/redux";
 import styled from "styled-components";
 import { UseCaseOption } from "./UseCaseOption";
