+++ apps/ledger-live-desktop/src/renderer/components/PostOnboardingHub/logic/useNavigateToPostOnboardingHubCallback.ts
@@ -2,7 +2,7 @@ import { useCallback } from "react";
 import { useSelector } from "LLD/hooks/redux";
 import { useNavigate } from "react-router";
 import { useFeature } from "@features/platform-feature-flags";
-import { isRecoverDisplayed } from "@ledgerhq/live-common/featureFlags/index";
+import { isRecoverDisplayed } from "@ledgerhq/live-common/recover/isRecoverDisplayed";
 import { useUpsellPath } from "@ledgerhq/live-common/hooks/recoverFeatureFlag";
 import { usePostOnboardingHubState } from "@ledgerhq/live-common/postOnboarding/hooks/index";
 import { hasBeenRedirectedToPostOnboardingSelector } from "~/renderer/reducers/settings";
