+++ apps/ledger-live-desktop/src/renderer/hooks/useAutoRedirectToPostOnboarding/useOpenPostOnboardingCallback.ts
@@ -2,7 +2,7 @@ import { DeviceModelId } from "@ledgerhq/types-devices";
 import { useCallback } from "react";
 import { useStartPostOnboardingCallback } from "@ledgerhq/live-common/postOnboarding/hooks/useStartPostOnboardingCallback";
 import { useFeature } from "@features/platform-feature-flags";
-import { isRecoverDisplayed } from "@ledgerhq/live-common/featureFlags/index";
+import { isRecoverDisplayed } from "@ledgerhq/live-common/recover/isRecoverDisplayed";
 import { useUpsellPath } from "@ledgerhq/live-common/hooks/recoverFeatureFlag";
 import { hasStartedLedgerRecoverFlowForPostOnboarding } from "LLD/features/FinishOnboarding/RecoverWidget/recoverPortfolioWidgetVisibility";
 import { getStoreValue } from "~/renderer/store";
