+++ apps/ledger-live-desktop/src/renderer/init.tsx
@@ -69,7 +69,7 @@ import {
 import { LiveConfig } from "@ledgerhq/live-config/LiveConfig";
 import { installLiveConfigProvider } from "~/firebase/remoteConfig";
 import { setAnalyticsFeatureFlagMethod } from "~/renderer/analytics/segment";
-import { FeatureId } from "@ledgerhq/types-live";
+import { FeatureId } from "@shared/feature-flags";
 import { initHistory } from "~/renderer/reducers/history";
 
 const rootNode = document.getElementById("react-root");
