+++ apps/ledger-live-desktop/src/renderer/screens/earn/index.tsx
@@ -1,7 +1,7 @@
 import {
   stakeProgramsToEarnParam,
   getEthDepositScreenSetting,
-} from "@ledgerhq/live-common/featureFlags/stakePrograms/index";
+} from "@ledgerhq/live-common/earn/stakePrograms/index";
 import { useFeature, useWalletFeaturesConfig } from "@features/platform-feature-flags";
 import { FEATURE_FLAGS_DEFAULTS } from "@shared/feature-flags";
 import {
