+++ libs/ledger-live-common/src/featureFlags/index.ts
@@ -8,5 +8,4 @@ export * from "./groupedFeatures";
 export * from "./FeatureFlagsContext";
 export * from "./useHasOverriddenFeatureFlags";
 export * from "./firebaseFeatureFlags";
-export * from "./helper";
 export * from "./walletFeaturesConfig";
